HvD: Fixing the interface for EAF_Stat. EAF_Stat reports the amount of diskspace in KB and used an int to hold that value. With the new Cascade machine in EMSL having a 2.6 PB /dtemp file system an int can no longer hold that value and EAF_Stat actually reported negative numbers. The fix is to upgrade the relevant function argument to long so we can deal with a billion times as much disk space.

+++ pario/eaf/eaf.c
@@ -506,7 +506,7 @@ int EAF_Delete(const char *fname)
  *
  * Return 0 on success, non-zero on failure.
  */
-int EAF_Stat(const char *path, int *avail_mb, char *fstype, int fslen)
+int EAF_Stat(const char *path, long *avail_mb, char *fstype, int fslen)
 {
     char dirname[PATH_MAX];
     stat_t statinfo;
@@ -516,7 +516,7 @@ int EAF_Stat(const char *path, int *avail_mb, char *fstype, int fslen)
     if ((rc = elio_stat(dirname, &statinfo))) return rc;
     if (fslen < 8) return EAF_ERR_TOO_SHORT;
 
-    *avail_mb = (int)(statinfo.avail>>10);
+    *avail_mb = (long)(statinfo.avail>>10);
     if (statinfo.fs == ELIO_UFS)
         strcpy(fstype, "UFS");
     else if (statinfo.fs == ELIO_PFS)
